feat: integrate Reanimated web hack functionality into AppRoot and MoreScreen
- Introduced a new custom hook, `useReanimatedWebHack`, to manage the Reanimated web hack state and toggle functionality.
- Updated `AppRoot` to utilize the new hook, ensuring the hack is enabled on component mount.
- Refactored `MoreScreen` to remove redundant hack logic and leverage the new hook for better state management.
- Enhanced user experience by providing a centralized way to control the Reanimated web hack across components.

These changes streamline the implementation and improve maintainability of the Reanimated web hack feature.
