+++ apps/playground/src/app/(tabs)/more.tsx
@@ -4,13 +4,14 @@ import {
     ListItem,
     ScreenWrapper,
     useThemePreferences,
-    useToast,
 } from '@siteed/design-system'
 import Constants from 'expo-constants'
 import { useRouter } from 'expo-router'
-import React, { useCallback, useEffect, useMemo, useState } from 'react'
+import React, { useMemo } from 'react'
 import { Image, Platform, StyleSheet, View } from 'react-native'
 import { Text } from 'react-native-paper'
+
+import { useReanimatedWebHack } from '../../hooks/useReanimatedWebHack'
 
 const getStyles = ({ theme }: { theme: AppTheme }) => {
     return StyleSheet.create({
@@ -40,43 +41,7 @@ export const MoreScreen = (_: MoreScreenProps) => {
     const { toggleDarkMode, darkMode, theme } = useThemePreferences()
     const styles = useMemo(() => getStyles({ theme }), [theme])
     const appVersion = Constants.expoConfig?.version
-    const [isHackEnabled, setIsHackEnabled] = useState(true)
-    const { show } = useToast()
-
-    useEffect(() => {
-        if (Platform.OS === 'web') {
-            handleHackToggle(true)
-        }
-    }, [])
-
-    const handleHackToggle = useCallback((value: boolean) => {
-        setIsHackEnabled(value)
-        if (Platform.OS === 'web') {
-            if (value) {
-                global._WORKLET = false
-                // @ts-expect-error
-                global._log = console.log
-                // @ts-expect-error
-                global._getAnimationTimestamp = () => performance.now()
-                show({
-                    type: 'success',
-                    iconVisible: true,
-                    message: 'Reanimated web hack enabled',
-                })
-            } else {
-                delete global._WORKLET
-                // @ts-expect-error
-                delete global._log
-                // @ts-expect-error
-                delete global._getAnimationTimestamp
-                show({
-                    type: 'warning',
-                    iconVisible: true,
-                    message: 'Reanimated web hack disabled',
-                })
-            }
-        }
-    }, [])
+    const { isHackEnabled, handleHackToggle } = useReanimatedWebHack()
 
     return (
         <ScreenWrapper withScrollView useInsets>
